+++ macos/desktop_multi_window/Package.swift
@@ -1,0 +1,25 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "desktop_multi_window",
+    platforms: [
+        .macOS("10.15")
+    ],
+    products: [
+        .library(name: "desktop-multi-window", targets: ["desktop_multi_window"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "desktop_multi_window",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ]
+        )
+    ]
+)